geolocation-system/backend/neural-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
